refactor: extract world book injection into WorldBookInjector
Move the world book prompt injection logic out of MessageBuilderService
into a standalone applyWorldBookInjections() so it can be reused by code
paths that run without a BuildContext (e.g. background isolates).
Pure extraction with no behavior change.
